# P02_SpringBoot/JPA/P07_CascadeOrphanRemoval/src/main/java/com/example/P07_CascadeOrphanRemoval/dto/DatabaseCascadeResult.java
package com.example.P07_CascadeOrphanRemoval.dto;

public record DatabaseCascadeResult(
        Long parentId,
        int childrenBeforeDirectSqlDelete,
        int parentRowsDeleted,
        int childrenAfterDirectSqlDelete,
        String mechanism
) {
}