# P02_SpringBoot/JPA/P07_CascadeOrphanRemoval/src/main/java/com/example/P07_CascadeOrphanRemoval/dto/NoCascadeRemoveResult.java
package com.example.P07_CascadeOrphanRemoval.dto;

public record NoCascadeRemoveResult(
        Long parentId,
        int childrenBeforeDelete,
        boolean deletionBlockedByForeignKey,
        boolean parentStillExists,
        String explanation
) {
}