# P02_SpringBoot/JPA/P07_CascadeOrphanRemoval/src/main/java/com/example/P07_CascadeOrphanRemoval/dto/OrphanRemovalResult.java
package com.example.P07_CascadeOrphanRemoval.dto;

public record OrphanRemovalResult(
        Long parentId,
        Long removedChildId,
        boolean parentExistsAfter,
        boolean removedChildExistsAfter,
        int remainingChildren
) {
}